% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregions_list.R
\docType{data}
\name{ecoregions_list}
\alias{ecoregions_list}
\title{Ecoregions List}
\format{
A data frame with 426 rows and 4 variables:
\describe{
\item{id}{An integer vector. The unique identifier for each ecoregion.}
\item{realm}{A character vector. The biogeographic realm.}
\item{major_habitat_type}{A character vector. The major habitat type for the ecoregion.}
\item{ecoregion}{A character vector. The specific name of the ecoregion.}
}
}
\source{
https://www.feow.org/ecoregions/list
}
\usage{
ecoregions_list
}
\description{
This dataset provides information about all FEOW ecoregions,
including the major habitat types and specific ecoregions.
}
\author{
The Nature Conservancy and World Wildlife Fund, Inc.
}
\keyword{datasets}
