% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_feow.R
\name{read_feow}
\alias{read_feow}
\title{Read FEOW HydroSheds data}
\source{
https://www.feow.org/ecoregions/list
}
\usage{
read_feow(overwrite = FALSE)
}
\value{
A spatial data frame containing the FEOW ecoregions polygons.
}
\description{
This function downloads and reads the FEOW ecoregions data from the
\code{brunomioto/feowR} repository.
}
\examples{
feow_data <- read_feow()
}
\author{
The Nature Conservancy and World Wildlife Fund, Inc.
}
