% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_feow.R
\name{read_feow}
\alias{read_feow}
\title{Read FEOW HydroSheds data
This function downloads and reads the FEOW HydroSheds data from the
\code{brunomioto/feowR} repository. It returns a spatial data frame containing
the FEOW HydroSheds catchment polygons.}
\usage{
read_feow()
}
\value{
A spatial data frame containing the FEOW HydroSheds catchment polygons.
}
\description{
Read FEOW HydroSheds data
This function downloads and reads the FEOW HydroSheds data from the
\code{brunomioto/feowR} repository. It returns a spatial data frame containing
the FEOW HydroSheds catchment polygons.
}
\examples{
feow_data <- read_feow()
}
